Updated UserController to restrict access to delete a task

+++ src/AppBundle/Controller/TaskController.php
@@ -128,12 +128,25 @@ class TaskController extends Controller
      */
     public function deleteTaskAction(Task $task)
     {
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($task);
-        $em->flush();
+        $author = $task->getUser();
+        $user = $this->getUser();
+        $roles = $user->getRoles();
+        $isAdmin = in_array('ROLE_ADMIN', $roles);
 
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
+        if (($author->getUsername() === 'Anonyme' && $isAdmin) || $user === $author) {
+            $em = $this->getDoctrine()->getManager();
+
+            $em->remove($task);
+            $em->flush();
+
+            $this->addFlash('success', 'La tâche a bien été supprimée !');
+
+            return $this->redirectToRoute('task_list');
+        }
+
+        $this->addFlash('error', 'Vous ne pouvez pas supprimer cette tâche !');
 
         return $this->redirectToRoute('task_list');
     }
+
 }
